+++ docs/Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.2.0 (Jul 2025)
+- Implemented Push changes to FCM and HMS [(#174)](https://github.com/wultra/mtoken-sdk-android/issues/174)
+
+## 2.1.0 (Jul 2025)
+
+- Added `mobileTokenData` to authorize request for passing customer-specific data ([documentation](Using-Operations-Service.md#Passing-Additional-Mobile-Token-Data))
+  - Available with PowerAuth server 1.10+
+  - Can be used for fraud detection systems (FDS) or other custom business logic
+
 ## 2.0.0 (Feb 2025)
 
 - SDK simplification and added `OIDC` activation feature [(#182)](https://github.com/wultra/mtoken-sdk-android/pull/182)